requirements.txt
arrow==0.10.0
certifi==2017.7.27.1
chardet==3.0.4
elasticsearch==5.4.0
idna==2.6
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
schedule==0.5.0
six==1.11.0
urllib3==1.22
